+++ portability-api-launcher/src/main/java/org/datatransferproject/api/launcher/TypeManager.java
@@ -38,7 +38,7 @@ public interface TypeManager {
    * Registers a model type. Extensions that introduce new model subtypes must registered them here
    * so they can be databound properly.
    *
-   * @param type the type to register.
+   * @param types the type to register.
    */
   void registerTypes(Class<?>... types);
 }
