+++ portability-api-launcher/src/main/java/org/datatransferproject/launcher/types/TypeManagerImpl.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.datatransferproject.launcher.impl;
+package org.datatransferproject.launcher.types;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
